Split test and add assertions
