+++ sbm-support-rewrite/testcode/maven-projects/multi-module-events/module-a/src/main/java/com/acme/SomeClass.java
@@ -1,0 +1,2 @@
+package com.acme;
+public class SomeClass {}
